Move schedule edit schemas out of +page.server.ts
Same fix as the reminder edit page: SvelteKit only allows
route-specific exports from `+page.server.ts`. The exported
`serviceScheduleFormSchema` and the inline `deferFormSchema` caused
the schedule edit page to 500. Both moved to
`web-v4/src/routes/vehicles/[id]/tasks/_schemas.ts`; the schedule
defer schema is exported as `scheduleDeferFormSchema` to
disambiguate it from the defer schema in the tasks index page
(which includes an `id` field).

Also renamed the `clear_defer` action to `clearDefer` in this
file, matching the rename already done in the tasks index page.
`DeferForm.svelte` already uses `?/clearDefer`.
